requirements.txt
torch
tqdm
pyyaml
numpy